Fixed coordinates calculation and wrote some tests
