+++ PixiEditorDotNetCore3/Models/Dialogs/CustomDialog.cs
@@ -7,7 +7,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows;
 
-namespace PixiEditorDotNetCore3.Models
+namespace PixiEditorDotNetCore3.Models.Dialogs
 {
     public abstract class CustomDialog : NotifyableObject
     {
