+++ PixiEditorDotNetCore3/Models/Dialogs/ImportFileDialog.cs
@@ -6,7 +6,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows;
 
-namespace PixiEditorDotNetCore3.Models
+namespace PixiEditorDotNetCore3.Models.Dialogs
 {
     class ImportFileDialog : CustomDialog
     {
