+++ PixiEditorDotNetCore3/Models/Position/Coordinates.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PixiEditorDotNetCore3.Models
+namespace PixiEditorDotNetCore3.Models.Position
 {
     public class Coordinates
     {
