+++ PixiEditorDotNetCore3/Models/Position/DoubleCords.cs
@@ -4,13 +4,13 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PixiEditorDotNetCore3.Models
+namespace PixiEditorDotNetCore3.Models.Position
 {
     public class DoubleCords
     {
         public Coordinates Coords1 { get; set; }
         public Coordinates Coords2 { get; set; }
-        
+
         public DoubleCords(Coordinates cords1, Coordinates cords2)
         {
             Coords1 = cords1;
