+++ PixiEditorDotNetCore3/Models/Tools/BitmapPixelChanges.cs
@@ -1,11 +1,12 @@
-﻿using System;
+﻿using PixiEditorDotNetCore3.Models.Position;
+using System;
 using System.Collections.Generic;
 using System.Text;
 using System.Windows.Media;
 
 namespace PixiEditorDotNetCore3.Models.Tools
 {
-     public struct BitmapPixelChanges
+    public struct BitmapPixelChanges
     {
         public Coordinates[] ChangedCoordinates { get; set; }
         public Color PixelsColor { get; set; }
