+++ PixiEditorDotNetCore3/Models/Tools/ShapeTool.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿using PixiEditorDotNetCore3.Models.Layers;
+using PixiEditorDotNetCore3.Models.Position;
+using System;
 using System.Collections.Generic;
 using System.Text;
 using System.Windows.Media;
