+++ PixiEditorDotNetCore3/Models/Tools/ToolManager.cs
