+++ PixiEditorDotNetCore3/Models/Tools/Tools/EarserTool.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿using PixiEditorDotNetCore3.Models.Layers;
+using PixiEditorDotNetCore3.Models.Position;
+using System;
 using System.Collections.Generic;
 using System.Text;
 using System.Windows.Media;
@@ -12,7 +14,7 @@ namespace PixiEditorDotNetCore3.Models.Tools.Tools
         public override BitmapPixelChanges Use(Layer layer, Coordinates startingCoords, Color color, int toolSize)
         {
             PenTool pen = new PenTool();
-            return pen.Draw(startingCoords, Colors.Transparent, toolSize);
+            return pen.Draw(startingCoords, System.Windows.Media.Colors.Transparent, toolSize);
         }
     }
 }
