+++ PixiEditorDotNetCore3/Models/Tools/Tools/LineTool.cs
@@ -1,4 +1,6 @@
-﻿using System.Windows.Media;
+﻿using PixiEditorDotNetCore3.Models.Layers;
+using PixiEditorDotNetCore3.Models.Position;
+using System.Windows.Media;
 using System.Windows.Media.Imaging;
 
 namespace PixiEditorDotNetCore3.Models.Tools.Tools
