+++ PixiEditorTests/WorkspaceTests/ImageGeneratorTests.cs
@@ -1,5 +1,5 @@
 ﻿using NUnit.Framework;
-using PixiEditorDotNetCore3.Models;
+using PixiEditorDotNetCore3.Models.Images;
 using System;
 using System.Collections.Generic;
 using System.Text;
